requirements.txt
pyspark==3.5.1
psycopg2-binary==2.9.9
lxml==5.1.0
pandas==2.2.1
